Remove unused query results logic
