+++ ui/src/logs/containers/LogsPage.tsx
@@ -11,7 +11,7 @@ import {fetchChunk} from 'src/logs/utils/fetchChunk'
 import {notify as notifyAction} from 'src/shared/actions/notifications'
 
 import {Greys} from 'src/reusable_ui/types'
-import QueryResults from 'src/logs/components/QueryResults'
+import HistogramResults from 'src/logs/components/HistogramResults'
 
 import {
   NOW,
@@ -737,10 +737,9 @@ class LogsPage extends Component<Props, State> {
 
     return (
       <div className="logs-viewer--graph-controls">
-        <QueryResults
+        <HistogramResults
           count={this.histogramTotal}
           queryCount={queryCount}
-          isInsideHistogram={true}
           searchStatus={searchStatus}
         />
         <TimeWindowDropdown
